API fixes.
The list isn't really a list; since previous versions of a list can sit
around in the current scope, forks are possible, which makes some classic
operations (filter, append, map) either awkward or not as meaningful.
However, some important operations are still meaningful, such as find and
fold.

I could disallow forking, but this looks interesting as is.