Snuck in 'root' field again, to facilitate appending.
I'm a nervous nelly when it comes to passing in an object, and "being
confident" that the same reference will come out the way I want it.
So, it's better to make the one you want from inside the function, return
that reference, and have it talk to whatever's waiting outside. That's why
we would like an append method for this structure (the final name matters
little.)

This commit doesn't introduce that method; it only poc's its plausibility.