List method 'foreach' only sees nodes with non-nil car fields.
This now allows for an implementation of an 'atop' method (a variation of
an append method.) Any "empty" nodes (which only serve as metatables for
future links in the chain) shouldn't count as list elements, even though
'atop' involves concatenating the (empty) root of one list onto the tip of
the other.

For example, when calculating the length of a list that has been placed
on top of another, the root of that list now has a cdr that leads to the top
of the other list, meaning that the old version of 'foreach' would
mistakenly count this node as a valid list element. So, we must skip it by
noting that it lacks a 'car' field.